requirements.txt
langchain-core
langchain_community 
langgraph 
langchain-groq 
langchain_google_genai
langchain_chroma
chromadb
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
python-dotenv
colorama
langserve
sse_starlette
uvicorn
gunicorn
fastapi
requests# Updated on 2025-08-16
# Updated on 2025-08-16
